streamlit_app/requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
tensorflow==2.17.0
joblib==1.3.2
google-cloud-storage==2.12.0
google-cloud-bigquery==3.12.0
scikit-learn==1.3.0